pyproject.toml
[tool.poetry]
name = "image-compress"
version = "0.1.0"
description = ""
authors = ["Lamar Marshall <lamarmarshall75@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pillow = "^7.1.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
